+++ altapay.php
@@ -14,6 +14,7 @@ require_once __DIR__ . '/vendor/autoload.php';
 
 class ALTAPAY extends PaymentModule
 {
+    const ALTAPAY_MANUAL_CAPTURE_REFUND_STATUS = 'no';
     public $url;
     public $captureStatus;
     public $username;
@@ -29,10 +30,10 @@ class ALTAPAY extends PaymentModule
     {
         $this->name = 'altapay';
         $this->tab = 'payments_gateways';
-        $this->version = '3.7.5';
+        $this->version = '3.7.6';
         $this->author = 'AltaPay A/S';
         $this->is_eu_compatible = 1;
-        $this->ps_versions_compliancy = ['min' => '1.6.0.1', 'max' => '8.1.3'];
+        $this->ps_versions_compliancy = ['min' => '1.6.0.1', 'max' => '8.1.5'];
         $this->currencies = true;
         $this->currencies_mode = 'checkbox';
         $this->bootstrap = true;
@@ -1073,7 +1074,10 @@ class ALTAPAY extends PaymentModule
                     $transaction = $response->Transactions[$latestTransKey];
                     updateTransactionIdForParentSubscription($orderID, $transaction->TransactionId);
                 }
-                if (markAsCaptured($paymentID, $this->getItemCaptureRefundQuantityCount($finalOrderLines))) {
+
+                $orderStatus = Configuration::get('manual_capture_payments_status');
+
+                if (markAsCaptured($paymentID, $this->getItemCaptureRefundQuantityCount($finalOrderLines)) && ($orderStatus !== self::ALTAPAY_MANUAL_CAPTURE_REFUND_STATUS)) {
                     $order->setCurrentState((int) Configuration::get('PS_OS_PAYMENT'));
                 }
                 saveOrderReconciliationIdentifier($orderID, $reconciliation_identifier);
@@ -1138,8 +1142,10 @@ class ALTAPAY extends PaymentModule
                     );
                     exit();
                 }
-
-                $order->setCurrentState((int) Configuration::get('PS_OS_REFUND'));
+                $refundStatus = Configuration::get('manual_refund_payments_status');
+                if ($refundStatus !== self::ALTAPAY_MANUAL_CAPTURE_REFUND_STATUS) {
+                    $order->setCurrentState((int) Configuration::get('PS_OS_REFUND'));
+                }
                 saveOrderReconciliationIdentifier($orderID, $reconciliation_identifier, 'refunded');
             } catch (Exception $e) {
                 $message = $e->getMessage();
@@ -1540,7 +1546,9 @@ class ALTAPAY extends PaymentModule
     public function renderForm()
     {
         $statuses = OrderState::getOrderStates($this->context->language->id);
-        $selectCaptureStatus = $selectAuthStatus = [];
+        $selectCaptureStatus = [];
+        $selectAuthStatus = [['id_option' => '0', 'name' => 'Select payment status']];
+
         foreach ($statuses as $status) {
             $selectCaptureStatus[] = ['key' => $status['id_order_state'], 'name' => $status['name']];
             $selectAuthStatus[] = ['id_option' => $status['id_order_state'], 'name' => $status['name']];
@@ -1625,6 +1633,46 @@ class ALTAPAY extends PaymentModule
                             'name' => 'name',
                         ],
                     ],
+                    [
+                        'type' => 'select',
+                        'label' => $this->l('Update status on manual capture'),
+                        'name' => 'manual_capture_payments_status',
+                        'required' => false,
+                        'options' => [
+                            'query' => [
+                                [
+                                    'id_option' => 'yes',
+                                    'name' => 'Yes',
+                                ],
+                                [
+                                    'id_option' => 'no',
+                                    'name' => 'No',
+                                ],
+                            ],
+                            'id' => 'id_option',
+                            'name' => 'name',
+                        ],
+                    ],
+                    [
+                        'type' => 'select',
+                        'label' => $this->l('Update status on manual refund'),
+                        'name' => 'manual_refund_payments_status',
+                        'required' => false,
+                        'options' => [
+                            'query' => [
+                                [
+                                    'id_option' => 'yes',
+                                    'name' => 'Yes',
+                                ],
+                                [
+                                    'id_option' => 'no',
+                                    'name' => 'No',
+                                ],
+                            ],
+                            'id' => 'id_option',
+                            'name' => 'name',
+                        ],
+                    ],
                 ],
                 'submit' => [
                     'title' => $this->l('Save'),
@@ -1664,6 +1712,9 @@ class ALTAPAY extends PaymentModule
      */
     public function getConfigFieldsValues()
     {
+        $enableCapture = Configuration::get('manual_capture_payments_status') ? Configuration::get('manual_capture_payments_status') : 'yes';
+        $enableRefund = Configuration::get('manual_refund_payments_status') ? Configuration::get('manual_refund_payments_status') : 'yes';
+
         return [
             'ALTAPAY_USERNAME' => Tools::getValue('ALTAPAY_USERNAME', Configuration::get('ALTAPAY_USERNAME')),
             'ALTAPAY_PASSWORD' => Tools::getValue('ALTAPAY_PASSWORD', Configuration::get('ALTAPAY_PASSWORD')),
@@ -1674,6 +1725,8 @@ class ALTAPAY extends PaymentModule
             'enable_fraud' => Tools::getValue('enable_fraud', Configuration::get('enable_fraud')),
             'enable_release_refund' => Tools::getValue('enable_release_refund', Configuration::get('enable_release_refund')),
             'authorized_payments_status' => Tools::getValue('authorized_payments_status', Configuration::get('authorized_payments_status')),
+            'manual_capture_payments_status' => Tools::getValue('manual_capture_payments_status', $enableCapture),
+            'manual_refund_payments_status' => Tools::getValue('manual_refund_payments_status', $enableRefund),
         ];
     }
 
@@ -1807,6 +1860,12 @@ class ALTAPAY extends PaymentModule
             }
             if (Tools::getValue('authorized_payments_status') !== '') {
                 Configuration::updateValue('authorized_payments_status', Tools::getValue('authorized_payments_status'));
+            }
+            if (Tools::getValue('manual_capture_payments_status') !== '') {
+                Configuration::updateValue('manual_capture_payments_status', Tools::getValue('manual_capture_payments_status'));
+            }
+            if (Tools::getValue('manual_refund_payments_status') !== '') {
+                Configuration::updateValue('manual_refund_payments_status', Tools::getValue('manual_refund_payments_status'));
             }
         }
         $this->Mhtml .= '<div class="alert alert-success"> ' . $this->l('Settings updated') . '</div>';